Run Uninstall Tests in isolation
This test results in Plugin deletion. If other tests exist in the same test-group, those would subsequently fail as there’s no Plugin to test.
