+++ .github/workflows/tests.yml
@@ -69,6 +69,7 @@ jobs:
           'EndToEnd/forms/general',
           'EndToEnd/forms/post-types',
           'EndToEnd/general/other',
+          'EndToEnd/general/uninstall',
           'EndToEnd/general/plugin-screens',
           'EndToEnd/integrations/divi-builder',
           'EndToEnd/integrations/divi-theme',
